+++ dca-archunit/src/test/java/dev/domaincentric/dca/archunit/fixtures/usecase/good/order/application/placeorder/PlaceOrderUseCase.java
@@ -5,8 +5,10 @@ import dev.domaincentric.dca.archunit.fixtures.usecase.good.order.domain.model.O
 import dev.domaincentric.dca.archunit.fixtures.usecase.good.order.domain.model.OrderId;
 import dev.domaincentric.dca.buildingblocks.hexagonal.port.out.DomainEventPublisher;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@Transactional
 public final class PlaceOrderUseCase implements PlaceOrderInputPort {
   private final OrderRepository orders;
   private final DomainEventPublisher events;